feat: Apply conditions for setting the second number
The second number can now only be set when certain checkpoints
in the dial are cleared, implying that a 360 degree turn is made. This
mimics the original masterlock mechanism. However, the algorithm is
not perfect yet. The user could turn more that 360 degrees to the right
rather that left to meet the conditions.