start: Starts google sign-in implementation
The user is now able to sign in with a google account via redirect to
goole OAuth page. As of now, the user stays logged in, but the app
is not aware of it after a refresh. This is expected because the app
calls the OAuth redirect callback again with no actual redirect when
refreshing. This will be fixed in the next commit.