fix: App does not retain the user logged in
The app now checks whether the user is logged in from the previous
session or not. This will prevent the user to appear logged out when
not since user data will still be available for the app after a refresh
anywhere in the app as long as the Header is there. The Header is
where the necessary checks are done.