fix: Updates get message function according to new firebase rules
The update function is still not ideal still it does not hide the passcode
and the token when retrieving the message. This is because firebase
database includes the whole message object fetching. The firebase
rules and structure will have to be updated in order for this to be
final. However, it looks like this is a firebase database limitation that
there is no conditional fetching based on query requests. Technically,
as long as you have the message id, your user id, and you are logged
in, you can get the whole message object.